2016 Day 22 in Python
The part 2 solution is NOT generic, because that would take
roughly until next Christmas to run. The inputs are limited to a
set of patterns. My input looked something like:

-------
|O...X|  O => Goal Destination
|.....|  X => Required Data
|..###|  # => Wall Node
|..._.|  _ => Empty Node
-------

From there, I worked out a formula that required the position of
the empty node and the leftmost wall node:

steps = y[empty] + 2 * (x[empty] - x[wall] + 1) + 5 * (width - 2)

The first term accounts for the vertical movement. The second
gets the empty node around the wall. The third is a series of
movements to move the required data to the goal.