Complete the 2016 AoC
This year was just as much fun as last year! The puzzles were a
bit repetitive (lots of md5, pathfinding, and assembunny), but
repetition and practice is the only way to get better. Actually
applying what I learned in my Game AI class was really cool, too.

I kind of regret not fighting for the leaderboard more often, but
I'm satisfied to finally get a position on the last day. Maybe
next year I'll go for the total top 100.