2017 Day 7
I finally finished this one. This solution handles a case that I
saw many other solutions did not: unbalanced nodes with fewer than
three children. As it turns out, no input like this is actually
given. That wasn't specified in the instructions, though, so I
wanted to make sure that I handled that case.