Fix 2017 directory structure
Move the 2017 puzzles to a Go directory and update the README.

This officialy wraps up AoC 2017. I liked the variety of the puzzles
this year more than last year, although I will always be partial to
a good pathfinding problem. It was nice to finally get experience
with Go, even if I haven't found it to be that great of a language
for this purpose.

Looking back at my last commit of last year, I acknowledge that I
did not compete for the leaderboard more... actually, I only tried
twice, and didn't quite make it either time. I guess that's a result
of working life; midnight is too late to stay up to code. And to
think that's when I used to get the most coding done.

Anyways, it's been fun, and I look forward to next year's AoC!